feat: return competitions with fixture EP; refactor jwt classes
